+++ src/components/Dashboard/Authentication/SignUp/Signup.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 
 import { connect } from "react-redux";
 import { NavLink, Link } from "react-router-dom";
-import { signup } from "../../../redux/actions";
+import { signup } from "../../../../redux/actions";
 
 import styled from "styled-components";
 import {
@@ -17,7 +17,7 @@ import {
   losTransitionTime,
   losBackground,
   privilege
-} from "../cssVariables";
+} from "../../cssVariables";
 
 const SignupPageContainer = styled.div`
   margin: 0 auto;
